% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommodityClassification.R
\name{getCommodityClassification}
\alias{getCommodityClassification}
\title{Get Commodity Classification}
\usage{
getCommodityClassification(cpcCodes)
}
\arguments{
\item{cpcCodes}{A character vector containing the CPC codes.}
}
\value{
A character vector whose ith element is the food classification
 corresponding to the passed CPC code in position i. If no valid cpc code is
 found, an NA is returned in this place.
}
\description{
This function classifies the cpc code in "Consumable, main", "Non-Consumable"
or "Consumable, residual".
}
\examples{
\dontrun{
getCommodityClassification(c("0111", "23110", "F0020"))
}

}

