% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFoodConsumption.R
\name{getFoodConsumption}
\alias{getFoodConsumption}
\title{Get Food Consumption}
\usage{
getFoodConsumption(yearRange)
}
\arguments{
\item{yearRange}{Character vector with the range of the years.}
}
\value{
The dataset with the food consumption.
}
\description{
This function pulls the food consumption data from the SWS
}

