% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNutrientConversionFactor.R
\name{getNutrientConversionFactors}
\alias{getNutrientConversionFactors}
\title{Get Nutrient Conversion Factor}
\usage{
getNutrientConversionFactors(country, item)
}
\arguments{
\item{country}{The m49 code of the "country" dimension.}

\item{item}{The cpc code of the "item" dimension.}
}
\value{
Codes
}
\description{
Get Nutrient Conversion Factor
}

